BC-11987 - Move sync module from infra to modules and rename to tsp-sync
