--- apps/server/src/infra/sync/index.ts
@@ -1,1 +1,0 @@
-export { TspSyncStrategy, TspUserInfo } from './strategy/tsp';
