+++ apps/server/src/modules/provisioning/service/tsp-provisioning.service.ts
@@ -1,10 +1,10 @@
 import { Logger } from '@infra/logger';
-import { TspUserInfo } from '@infra/sync/strategy/tsp/';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { Account, AccountSave, AccountService } from '@modules/account';
 import { Class, ClassFactory, ClassService, ClassSourceOptions } from '@modules/class';
 import { RoleName, RoleService } from '@modules/role';
 import { School, SchoolService } from '@modules/school';
+import { TspUserInfo } from '@modules/tsp-sync';
 import { Consent, ParentConsent, UserConsent, UserDo, UserService } from '@modules/user';
 import { Injectable } from '@nestjs/common';
 import { TypeGuard } from '@shared/common/guards';
