+++ apps/server/src/modules/tsp-sync/index.ts
@@ -1,0 +1,3 @@
+export { SyncConsole } from './console/sync.console';
+export { TspSyncStrategy, TspUserInfo } from './strategy/tsp';
+export { TspSyncModule } from './tsp-sync.module';
