+++ apps/server/src/modules/tsp-sync/strategy/tsp/tsp-oauth-data.mapper.ts
